Tower and Second Observatory
Added code to expand searchable locations to include the vertical component as well, so now objects can be hidden in any location, at any level. Set up a second 'observatory' as a result of this. Also took advantage of the new tower icon which I had made, but was not using.