Words of Power and more town dialogues
Experimented with words of power. They are working, but it uses up the last bits of code. They could be part of a nice set of quests, akin to the originals mantras, but the cost might be too high. Something else might need to be cut to fit them in.
Also nearly finished up the new hamlets base dialogues. Fixed the swamp town a bit, to look more natural.